+++ core/utils/cli_options.ts
@@ -6,7 +6,7 @@ export function getOptionsFromCli(
   options: DeepPartial<SiteOptions>,
 ): DeepPartial<SiteOptions> {
   const cli = parseArgs(Deno.args, {
-    string: ["src", "dest", "location", "port"],
+    string: ["src", "dest", "location", "port", "hostname"],
     boolean: ["serve", "open"],
     alias: { serve: "s", port: "p", open: "o" },
     ["--"]: true,
@@ -22,8 +22,9 @@ export function getOptionsFromCli(
 
   const serveMode = cli.serve || cli._[0] === "cms";
 
-  // Detect location and port
+  // Detect location, hostname and port
   let port: number;
+  let hostname: string;
   let location: URL;
 
   if (serveMode) {
@@ -43,6 +44,16 @@ export function getOptionsFromCli(
       port = location.protocol === "https:" ? 443 : 3000;
       location.port = port.toString();
     }
+
+    if (cli.hostname) {
+      hostname = cli.hostname;
+      location.hostname = hostname;
+    } else if (options.server?.hostname && !cli.location) {
+      hostname = options.server.hostname;
+      location.hostname = hostname;
+    } else {
+      hostname = location.hostname;
+    }
   } else {
     location = cli.location
       ? new URL(cli.location)
@@ -56,11 +67,19 @@ export function getOptionsFromCli(
     } else {
       port = location.protocol === "https:" ? 443 : 80;
     }
+
+    if (cli.hostname) {
+      hostname = cli.hostname;
+      location.hostname = hostname;
+    } else {
+      hostname = location.hostname;
+    }
   }
 
   options.location = location;
   options.server ||= {};
   options.server.port = port;
+  options.server.hostname = hostname;
 
   if (cli.open) {
     options.server.open = cli.open;