+++ core/utils/net.ts
@@ -12,3 +12,23 @@ export function localIp(): string | undefined {
     return undefined;
   }
 }
+
+export async function openBrowser(url: string): Promise<void> {
+  const commands: Record<typeof Deno.build.os, string> = {
+    darwin: "open",
+    linux: "xdg-open",
+    freebsd: "xdg-open",
+    netbsd: "xdg-open",
+    aix: "xdg-open",
+    solaris: "xdg-open",
+    illumos: "xdg-open",
+    windows: "explorer",
+    android: "xdg-open",
+  };
+
+  await new Deno.Command(commands[Deno.build.os], {
+    args: [url],
+    stdout: "inherit",
+    stderr: "inherit",
+  }).output();
+}